Create Ambassador Set.json
The new weapon the Ambassador is not on the site right now. So I thought I'd try contributing it here. (I don't use github often, so bare with me if I messed up something. And also don't know if it's already in the pipeline.)

The .json text will need a "thumb", "uniq_name" and "ID". I left them blank (" ").